pyproject.toml
[tool.poetry]
name = "fast-api-kit"
version = "0.1.0"
description = ""
authors = ["Loc Nguyen <V_NGUL509@kellyservices.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
fastapi = "^0.101.1"
aiohttp = "^3.8.5"
sqlalchemy = "^2.0.20"
pydantic-settings = "^2.0.3"
dependency-injector = "^4.41.0"
email-validator = "^2.0.0.post2"
psycopg2 = "^2.9.7"
loguru = "^0.7.1"
azure-identity = "^1.14.0"
msgraph-core = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
